refactor: streamLine 함수 분리
